refactor: split mcp server metadata and entrypoint
